% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random-words-and-sentences.R
\name{get_languages}
\alias{get_languages}
\title{Get shortcodes for all implemented languages.}
\usage{
get_languages()
}
\description{
Get shortcodes for all implemented languages.
}
\examples{
get_languages()
}

