% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random-words-and-sentences.R
\name{load_languages}
\alias{load_languages}
\title{Load language reference lists into memory}
\usage{
load_languages(...)
}
\arguments{
\item{...}{Language shortcodes as character vector(s), e.g. \code{("en",
"de")}. See \code{\link{get_languages}} for available shortcodes.}
}
\description{
By default an English wordlist is loaded. To show available languages use
\code{\link{get_languages}}.
}
\examples{
load_languages("de", "nl")
}

