% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random-words-and-sentences.R
\name{print.wordlist}
\alias{print.wordlist}
\title{Print method for wordlist object}
\usage{
\method{print}{wordlist}(x)
}
\description{
Object is created in zzz.R.
}
\keyword{internal}

