% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random-words-and-sentences.R
\name{sentence}
\alias{sentence}
\title{Random sentence with n words}
\usage{
sentence(w, lang = "en", maxchar = Inf)
}
\arguments{
\item{w}{Number of words in sentence.}

\item{maxchar}{Maximal number of characters per sentence. Note that whole
words (not part of words) are excluded if the maximal number
is exceeded.}
}
\value{
a string with n words (if length is not constrained)
}
\description{
Random sentence with n words
}
\examples{
sentence(5)            # random sentence with 5 words
sentence(5, max=20)   # random sentence cut off after 20 chars
}
\keyword{internal}

