% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random-words-and-sentences.R
\name{sentences}
\alias{sentences}
\title{Random sentences}
\usage{
sentences(w, s = 1, lang = "en", maxchar = Inf)
}
\arguments{
\item{w}{Number of words per sentence. If more than one value
lengths is randomly drawn from \code{w}.}

\item{s}{Number of sentences to generate (integer).}

\item{maxchar}{Maximal number of characters per sentence. Note that whole
words (not part of words) are excluded if the maximal number
is exceeded.}
}
\value{
A vector of random sentences.
}
\description{
Create random sentences with a given or random number of words.
}
\examples{
# one sentence
sentences(5)            # random sentence with 5 words
sentences(5, max=30)    # cut off after 20 chars
sentences(3:5)          # between 3 and 5 words

# many sentences
sentences(3, 2)         # 2 sentences each 3 words long
sentences(3:5, 2)       # 2 sentences each between 3 and 5 words long
}

