% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/words.R
\docType{package}
\name{words-package}
\alias{words-package}
\title{\pkg{words} creates random words or sentences}
\description{
Sometimes random words or sentences are needed as dummys or for simulation
purposes. The functions in this package create random words or sentences in
several languages. Note that currently the senteces are a random
concatenation of words without any meaning.
}
\details{
In the future reference lists containing the most frequent ngrams may be
added to have sentences that make some sense.

Word lists are loaded once and are stored in the invisible object \code{.words}.
}
\author{
Mark Heckmann
}
\keyword{package}
\keyword{words}

