% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random-words-and-sentences.R
\name{words}
\alias{words}
\title{Random words}
\usage{
words(n = 1, lang = "en")
}
\arguments{
\item{n}{Number of words to be generated (integer).}

\item{lang}{Language code (default \code{"en"}). See \code{\link{get_languages}} for available
language codes.}
}
\value{
A vector of random words.
}
\description{
\code{words} generates a vector of random words taken from a reference list
in the prompted language.
}
\examples{
words(10)  # 10 random words
}

