% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random-words-and-sentences.R
\name{get_languages}
\alias{get_languages}
\title{get languages for which wordlists are available}
\usage{
get_languages()
}
\description{
get languages for which wordlists are available
}

