% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/random-words-and-sentences.R
\name{load_languages}
\alias{load_languages}
\title{Load language reference lists into memory}
\usage{
load_languages(...)
}
\description{
By default english is loaded. To show the available languages use
\code{\link{get_languages}}.
}

