% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{words}
\alias{words}
\title{List of english words}
\format{
A data frame with 175393 rows and 2 variables:
\describe{
\item{word}{an english word}
\item{word_length}{number of letters in the word}
...
}
}
\source{
\url{https://www.wordgamedictionary.com/word-lists/}
}
\usage{
words
}
\description{
A dataset containing a list of english words and their length.
}
\keyword{datasets}
